src/main/ngTime/src/app/+export/shared/export-link/export-link.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { ExportLinkComponent } from './export-link.component';

describe('Component: ExportLink', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [ExportLinkComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([ExportLinkComponent],
      (component: ExportLinkComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(ExportLinkComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(ExportLinkComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-export-link></app-export-link>
  `,
  directives: [ExportLinkComponent]
})
class ExportLinkComponentTestController {
}

